Update README.md
Updated README and Project Documentation

This commit includes significant updates to the README and project documentation. The changes include:

1. Updated the project's domain from rayice.com to digitalcloud.no in all relevant sections of the README.
2. Added a 'Future Plans' section to the README to provide information about planned updates and improvements to the project.
3. Expanded the descriptions in the README to provide more detailed information about the project, its features, and its usage.

These changes were made to provide clearer and more up-to-date information about the project to users and contributors. They reflect the ongoing efforts to modernize the project and make it more accessible and useful to a wider audience.